+++ src/core/lang/sv/mod.rs
@@ -17,6 +17,7 @@
 
 pub mod dst;
 pub mod error;
+pub mod format;
 pub mod primaryunit;
 pub mod symbols;
 pub mod token;