+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 ### Features
 - adds initial ability to use `--json` for verilog and systemverilog modules with `orbit get` command
 - adds global flag `--sync` to synchronize channels if they are stored on the internet as well as locally on the user's system
+- adds formatting options for verilog and sv languages by adding new table entry in configuration files: "systemverilog-format"
 
 ### Changes
 - formats the ip's manifest for writing it to the output index path of its channel during `orbit publish`